requirements.txt
anyio==3.6.1
asyncpg==0.26.0
black==22.10.0
click==8.1.3
fastapi==0.85.0
greenlet==1.1.3
h11==0.14.0
h2==4.1.0
hpack==4.0.0
hypercorn==0.14.3
hyperframe==6.0.1
idna==3.4
mypy-extensions==0.4.3
pathspec==0.10.1
platformdirs==2.5.2
priority==2.0.0
pydantic==1.10.2
PyYAML==6.0
sniffio==1.3.0
SQLAlchemy==1.4.41
starlette==0.20.4
toml==0.10.2
tomli==2.0.1
typing_extensions==4.4.0
wsproto==1.2.0
